+++ src/index.js
@@ -49,6 +49,7 @@ import {
     BACK_Z_INDEX,
 } from './blobs';
 import { ULABEL_VERSION } from './version';
+import { BrushToolboxItem } from '../build/toolbox';
 
 jQuery.fn.outer_html = function () {
     return jQuery('<div />').append(this.eq(0).clone()).html();
@@ -1362,7 +1363,6 @@ export class ULabel {
      */
     get_current_subtask() {
         return this.subtasks[this.get_current_subtask_key()];
-        
     }
 
     readjust_subtask_opacities() {
@@ -2783,6 +2783,8 @@ export class ULabel {
         // Try and switch to polygon mode if not already in it
         if (!is_in_polygon_mode) {
             is_in_polygon_mode = this.set_and_update_annotation_mode("polygon");
+            $("#brush-mode").removeClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
+            $("#erase-mode").removeClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
         }
         // If we're in polygon mode, toggle brush mode
         if (is_in_polygon_mode) {
@@ -2807,30 +2809,45 @@ export class ULabel {
                 let gmx = this.get_global_mouse_x(mouse_event);
                 let gmy = this.get_global_mouse_y(mouse_event);
                 this.create_brush_circle(gmx, gmy);
+                $("#brush-mode").addClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
             } else {
                 this.destroy_brush_circle();
+                $("#brush-mode").removeClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
             }
         }
     }
 
     toggle_erase_mode(mouse_event) {
-        const current_subtask = this.get_current_subtask_key();
+        const current_subtask = this.get_current_subtask();
         // If not in brush mode, turn it on
-        if (!this.subtasks[current_subtask]["state"]["is_in_brush_mode"]) {
+        if (!current_subtask["state"]["is_in_brush_mode"]) {
             this.toggle_brush_mode(mouse_event);
         }
 
         // Toggle erase mode
-        this.subtasks[current_subtask]["state"]["is_in_erase_mode"] = !this.subtasks[current_subtask]["state"]["is_in_erase_mode"];
+        if (current_subtask["state"]["is_in_erase_mode"]) {
+            $("#erase-mode").removeClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
+            // "Erase mode" is a subset of "brush mode"
+            if (current_subtask["state"]["is_in_brush_mode"]) {
+                $("#brush-mode").addClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
+            }
+        } else {
+            $("#erase-mode").addClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
+            $("#brush-mode").removeClass(BrushToolboxItem.BRUSH_BTN_ACTIVE_CLS);
+        }
+        current_subtask["state"]["is_in_erase_mode"] = !current_subtask["state"]["is_in_erase_mode"];
 
         // Update brush circle color
         const brush_circle_id = "brush_circle";
         $("#" + brush_circle_id).css({
-            "background-color": this.subtasks[current_subtask]["state"]["is_in_erase_mode"] ? "red" : "white",
+            "background-color": current_subtask["state"]["is_in_erase_mode"] ? "red" : "white",
         });
 
         // When turning off erase mode, also turn off brush mode
-        if (this.subtasks[current_subtask]["state"]["is_in_brush_mode"] && !this.subtasks[current_subtask]["state"]["is_in_erase_mode"]) {
+        if (
+            current_subtask["state"]["is_in_brush_mode"] &&
+            !current_subtask["state"]["is_in_erase_mode"]
+        ) {
             this.toggle_brush_mode();
         }
     }